+++ src/lodi-client/stream_feed.h
@@ -1,7 +1,7 @@
 #ifndef COSC522_LODI_PRINT_FEED_H
 #define COSC522_LODI_PRINT_FEED_H
 
-int startStreamFeed(unsigned int userId, unsigned int timestamp, unsigned int digitalsig);
+int startStreamFeed(unsigned int userId, unsigned long timestamp, unsigned long digitalsig);
 
 void stopStreamFeed(int pid);
